refactor(bayn): unify Alpaca HTTP resource ownership
